# src/main/java/de/ait/todo/dto/NewUserDto.java
package de.ait.todo.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;


@Data
@AllArgsConstructor
@NoArgsConstructor
public class NewUserDto {
    private String email;
    private String password;
}