# src/main/java/de/ait/todo/dto/TasksPage.java
package de.ait.todo.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;


@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
@Schema(description = "Страница с задачами")
public class TasksPage {

    @Schema(description = "Список задач")
    private List<TaskDto> tasks;
}